--- a/test-fixtures/windows.txt
+++ b/test-fixtures/windows.txt
@@ -1,0 +1,3 @@
+goroutine 1 [running]:
+runtime/pprof.writeGoroutineStacks(0x14e5940, 0xc0000abc50, 0x101b8a5, 0xc000333d20)
+	C:/Users/felixge/pprof.go:693 +0x9f