Seemss like this also made the code a little faster :)

name             old time/op    new time/op    delta
Gostackparse-12    26.9µs ± 6%    25.3µs ± 1%   -5.97%  (p=0.000 n=10+10)

name             old MiB/s      new MiB/s      delta
Gostackparse-12       295 ± 5%       314 ± 1%   +6.27%  (p=0.000 n=10+10)

name             old alloc/op   new alloc/op   delta
Gostackparse-12    17.2kB ± 0%    17.2kB ± 0%   -0.19%  (p=0.000 n=10+10)

name             old allocs/op  new allocs/op  delta
Gostackparse-12       306 ± 0%       248 ± 0%  -18.95%  (p=0.000 n=10+10)